requirements.txt
Django==3.2.5
django-seed==0.2.2
djangorestframework==3.12.4
Faker==8.10.3
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
